+++ dwara-db/src/data/sql/dwara_update_2_1_8.sql
@@ -122,4 +122,15 @@ INSERT INTO `action_artifactclass_flow` (`action_id`, `artifactclass_id`, `flow_
 ('ingest', 'photo-edit-priv2', 'photo-proxy-flow', 0);
 
 
-SET foreign_key_checks = 1; 
+UPDATE `copy` SET `location_id` = 'lto-room';
+-- ?? UPDATE `volume` set `location_id` = 'lto-room' where `location_id` is not null;
+DELETE FROM `location` WHERE (`id` != 'lto-room');
+
+UPDATE `location` SET `description` = 'SK Office - 1st' WHERE (`id` = 'lto-room');
+INSERT INTO `location` (`id`, `default`, `description`) VALUES ('t-block2', b'0', 'T Block - 2nd');
+INSERT INTO `location` (`id`, `default`, `description`) VALUES ('t-block3', b'0', 'T Block - 3rd');
+
+UPDATE `copy` SET `location_id` = 't-block2' WHERE (`id` = 2);
+UPDATE `copy` SET `location_id` = 't-block3' WHERE (`id` = 3);
+
+SET foreign_key_checks = 1;